chore(test/ui): moved CreatFBTTest and BasicFBTOperationTest to org.eclipse.fordiac.ide.test.ui.fbtype package
