+++ tests/org.eclipse.fordiac.ide.test.ui/src/org/eclipse/fordiac/ide/test/ui/fbtype/BasicFBTOperationTest.java
@@ -11,7 +11,7 @@
  *
  *   Prashantkumar Khatri - initial API and implementation and/or initial documentation
  *******************************************************************************/
-package org.eclipse.fordiac.ide.test.ui.BasicFBTOperation;
+package org.eclipse.fordiac.ide.test.ui.fbtype;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
